-enh Made the tests compatible with the strong enums

+++ Roomba/src/UART/UART.cpp
@@ -8,7 +8,7 @@ UART::UART(const UARTSettings Settings)
 {
 
 		std::cout << "New instance of the UART object" << '\n';
-		std::cout << "Uartsettings Path: " << Settings.DevicePath << " Baudrate: " << Settings.baudrate << '\n';
+		std::cout << "Uartsettings Path: " << Settings.DevicePath << " Baudrate: " << static_cast<uint32_t>(Settings.baudrate) << '\n';
 
         uint32_t BaudrateVal = static_cast<uint32_t>(Settings.baudrate);
